move 'RequestedAppointmentDetailsPage' and unit test to 'pages' directory
