+++ src/applications/vaos/appointment-list/pages/RequestedAppointmentDetailsPage/RequestedAppointmentDetailsPage.jsx
@@ -1,23 +1,26 @@
-import React, { useEffect } from 'react';
-import { useParams } from 'react-router-dom';
-import { shallowEqual, useDispatch, useSelector } from 'react-redux';
 import { VaAlert } from '@department-of-veterans-affairs/component-library/dist/react-bindings';
-import BackLink from '../../components/BackLink';
-import { scrollAndFocus } from '../../utils/scrollAndFocus';
-import VAFacilityLocation from '../../components/VAFacilityLocation';
-import { selectRequestedAppointmentDetails } from '../redux/selectors';
-import { selectFeatureBreadcrumbUrlUpdate } from '../../redux/selectors';
-import ErrorMessage from '../../components/ErrorMessage';
-import PageLayout from './PageLayout';
-import FullWidthLayout from '../../components/FullWidthLayout';
-import { closeCancelAppointment, fetchRequestDetails } from '../redux/actions';
-import CancelWarningPage from './cancel/CancelWarningPage';
-import CancelConfirmationPage from './cancel/CancelConfirmationPage';
-import { FETCH_STATUS } from '../../utils/constants';
-import FacilityAddress from '../../components/FacilityAddress';
-import FacilityPhone from '../../components/FacilityPhone';
-import VARequestLayout from '../../components/layout/VARequestLayout';
-import CCRequestLayout from '../../components/layout/CCRequestLayout';
+import React, { useEffect } from 'react';
+import { shallowEqual, useDispatch, useSelector } from 'react-redux';
+import { useParams } from 'react-router-dom';
+import BackLink from '../../../components/BackLink';
+import ErrorMessage from '../../../components/ErrorMessage';
+import FacilityAddress from '../../../components/FacilityAddress';
+import FacilityPhone from '../../../components/FacilityPhone';
+import FullWidthLayout from '../../../components/FullWidthLayout';
+import CCRequestLayout from '../../../components/layout/CCRequestLayout';
+import VARequestLayout from '../../../components/layout/VARequestLayout';
+import VAFacilityLocation from '../../../components/VAFacilityLocation';
+import { selectFeatureBreadcrumbUrlUpdate } from '../../../redux/selectors';
+import { FETCH_STATUS } from '../../../utils/constants';
+import { scrollAndFocus } from '../../../utils/scrollAndFocus';
+import CancelConfirmationPage from '../../components/cancel/CancelConfirmationPage';
+import CancelWarningPage from '../../components/cancel/CancelWarningPage';
+import PageLayout from '../../components/PageLayout';
+import {
+  closeCancelAppointment,
+  fetchRequestDetails,
+} from '../../redux/actions';
+import { selectRequestedAppointmentDetails } from '../../redux/selectors';
 
 export default function RequestedAppointmentDetailsPage() {
   const { id } = useParams();
